Use a uint16_t for QUIC port values.
This reduces narrowing conversions.

Bug: 1216696
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2958228
Auto-Submit: Peter Kasting <pkasting@chromium.org>
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Commit-Queue: Victor Vasiliev <vasilvv@chromium.org>
Cr-Commit-Position: refs/heads/master@{#892240}

+++ net/tools/quic/quic_transport_simple_server.h
@@ -27,7 +27,7 @@ class QuicTransportSimpleServer {
  public:
   using ReadErrorCallback = base::OnceCallback<void(int)>;
 
-  QuicTransportSimpleServer(int port,
+  QuicTransportSimpleServer(uint16_t port,
                             std::vector<url::Origin> accepted_origins,
                             std::unique_ptr<quic::ProofSource> proof_source);
   ~QuicTransportSimpleServer();
@@ -50,7 +50,7 @@ class QuicTransportSimpleServer {
   // Passes the most recently read packet into the dispatcher.
   void ProcessReadPacket(int result);
 
-  const int port_;
+  const uint16_t port_;
 
   ReadErrorCallback read_error_callback_;
